Move waypoint module into its own package
